+++ contrib/jenkins_tests/cov.sh
@@ -4,15 +4,20 @@ source $(dirname $0)/globals.sh
 
 check_filter "Checking for coverity ..." "on"
 
+# This unit requires module so check for existence
+if [ $(command -v module >/dev/null 2>&1 || echo $?) ]; then
+	echo "[SKIP] module tool does not exist"
+	exit 0
+fi
+module load tools/cov
+
 cd $WORKSPACE
 
 rm -rf $cov_dir
 mkdir -p $cov_dir
 cd $cov_dir
 
-cov_exclude_file_list="tests src/vma/lwip"
-
-module load tools/cov
+cov_exclude_file_list="tests"
 
 cov_build_id="cov_build_${BUILD_NUMBER}"
 cov_build="$cov_dir/$cov_build_id"
@@ -29,7 +34,7 @@ for excl in $cov_exclude_file_list; do
     sleep 1
 done
 
-eval "cov-analyze --enable-fnptr --fnptr-models --all --paths 20000 --dir $cov_build --config ${WORKSPACE}/coverity_vma_config.xml"
+eval "cov-analyze --enable-fnptr --fnptr-models --all --paths 20000 --dir $cov_build"
 rc=$(($rc+$?))
 
 set -eE
